requirements.txt
requests==2.26.0
websocket-client==1.3.2
msgpack==1.0.3